+++ book-exchange-backend-mvc/src/models/bookImages.services.ts
@@ -3,22 +3,22 @@ import {
   CreateBookImageRequest,
   UpdateBookImageRequest,
 } from "../types/base/bookImages.types";
-
+import { PrismaTransactionClient } from "@src/types/utils/primaAlias.types";
 export const craeteBookImage = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   data: CreateBookImageRequest
 ) => {
   const bookImage = await prisma.bookImages.create({ data });
   return bookImage;
 };
 
-export const getBookImageById = async (prisma: PrismaClient, id: number) => {
+export const getBookImageById = async (prisma: PrismaTransactionClient, id: number) => {
   const bookImage = await prisma.bookImages.findUnique({ where: { id } });
   return bookImage;
 };
 
 export const updateBookImage = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   id: number,
   data: UpdateBookImageRequest
 ) => {
@@ -26,7 +26,7 @@ export const updateBookImage = async (
   return bookImage;
 };
 
-export const deleteBookImage = async (prisma: PrismaClient, id: number) => {
+export const deleteBookImage = async (prisma: PrismaTransactionClient, id: number) => {
   const bookImage = await prisma.bookImages.delete({ where: { id } });
   return bookImage;
 };