+++ book-exchange-backend-mvc/src/models/books.services.ts
@@ -1,24 +1,25 @@
 import { PrismaClient } from "@prisma/client";
+import { PrismaTransactionClient } from "@src/types/utils/primaAlias.types";
 import {
   CreateBookRequest,
   UpdateBookRequest,
 } from "../types/base/books.types";
 
 export const createBook = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   data: CreateBookRequest
 ) => {
   const book = await prisma.books.create({ data });
   return book;
 };
 
-export const getBookById = async (prisma: PrismaClient, id: number) => {
+export const getBookById = async (prisma: PrismaTransactionClient, id: number) => {
   const book = await prisma.books.findUnique({ where: { id } });
   return book;
 };
 
 export const updateBook = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   id: number,
   data: UpdateBookRequest
 ) => {
@@ -26,7 +27,7 @@ export const updateBook = async (
   return book;
 };
 
-export const deleteBook = async (prisma: PrismaClient, id: number) => {
+export const deleteBook = async (prisma: PrismaTransactionClient, id: number) => {
   const book = await prisma.books.delete({ where: { id } });
   return book;
 };