+++ book-exchange-backend-mvc/src/models/genres.services.ts
@@ -1,24 +1,25 @@
 import { PrismaClient } from "@prisma/client";
+import { PrismaTransactionClient } from "@src/types/utils/primaAlias.types";
 import {
   CreateGenreRequest,
   UpdateGenreRequest,
 } from "../types/base/genres.types";
 
 export const createGenre = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   data: CreateGenreRequest
 ) => {
   const genre = await prisma.genres.create({ data });
   return genre;
 };
 
-export const getGenreById = async (prisma: PrismaClient, id: number) => {
+export const getGenreById = async (prisma: PrismaTransactionClient, id: number) => {
   const genre = await prisma.genres.findUnique({ where: { id } });
   return genre;
 };
 
 export const updateGenre = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   id: number,
   data: UpdateGenreRequest
 ) => {
@@ -26,7 +27,7 @@ export const updateGenre = async (
   return genre;
 };
 
-export const deleteGenre = async (prisma: PrismaClient, id: number) => {
+export const deleteGenre = async (prisma: PrismaTransactionClient, id: number) => {
   const genre = await prisma.genres.delete({ where: { id } });
   return genre;
 };