+++ book-exchange-backend-mvc/src/models/logSettings.services.ts
@@ -1,21 +1,21 @@
 import { PrismaClient } from "@prisma/client";
 import { CreateLogSettingRequest, UpdateLogSettingRequest } from "@src/types";
-
+import { PrismaTransactionClient } from "@src/types/utils/primaAlias.types";
 export const createLogSettings = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   data: CreateLogSettingRequest
 ) => {
   const logSetting = await prisma.logSettings.create({ data });
   return logSetting;
 };
 
-export const getLogSettingById = async (prisma: PrismaClient, id: number) => {
+export const getLogSettingById = async (prisma: PrismaTransactionClient, id: number) => {
   const logSetting = await prisma.logSettings.findUnique({ where: { id } });
   return logSetting;
 };
 
 export const updateLogSetting = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   id: number,
   data: UpdateLogSettingRequest
 ) => {
@@ -23,6 +23,6 @@ export const updateLogSetting = async (
   return logSetting;
 };
 
-export const deleteLogSetting = async (prisma: PrismaClient, id: number) => {
+export const deleteLogSetting = async (prisma: PrismaTransactionClient, id: number) => {
   const logSetting = await prisma.logSettings.delete({ where: { id } });
 };