+++ book-exchange-backend-mvc/src/models/logs.services.ts
@@ -1,22 +1,23 @@
 import { PrismaClient } from "@prisma/client";
+import { PrismaTransactionClient } from "@src/types/utils/primaAlias.types";
 import prisma from "@src/services/prismaService";
 import { CreateLogRequest, UpdateLogRequest } from "@src/types";
 
 export const createLog = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   data: CreateLogRequest
 ) => {
   const log = await prisma.logs.create({ data });
   return log;
 };
 
-export const getLogById = async (prisma: PrismaClient, id: number) => {
+export const getLogById = async (prisma: PrismaTransactionClient, id: number) => {
   const log = await prisma.logs.findUnique({ where: { id } });
   return log;
 };
 
 export const updateLog = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   id: number,
   data: UpdateLogRequest
 ) => {
@@ -24,7 +25,7 @@ export const updateLog = async (
   return log;
 };
 
-export const deleteLog = async (params: PrismaClient, id: number) => {
+export const deleteLog = async (params: PrismaTransactionClient, id: number) => {
   const log = await prisma.logs.delete({ where: { id } });
   return log;
 };