+++ book-exchange-backend-mvc/src/models/notifications.services.ts
@@ -1,24 +1,26 @@
 import { PrismaClient } from "@prisma/client";
+import { PrismaTransactionClient } from "@src/types/utils/primaAlias.types";
 import {
   CreateNotificationRequest,
   UpdateNotificationRequest,
 } from "@src/types";
 
+
 export const createNotification = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   data: CreateNotificationRequest
 ) => {
   const notification = await prisma.notifications.create({ data });
   return notification;
 };
 
-export const getNotificationById = async (prisma: PrismaClient, id: number) => {
+export const getNotificationById = async (prisma: PrismaTransactionClient, id: number) => {
   const notification = await prisma.notifications.findUnique({ where: { id } });
   return notification;
 };
 
 export const updateNotificaiton = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   id: number,
   data: UpdateNotificationRequest
 ) => {
@@ -29,7 +31,7 @@ export const updateNotificaiton = async (
   return notification;
 };
 
-export const deleteNotification = async (prisma: PrismaClient, id: number) => {
+export const deleteNotification = async (prisma: PrismaTransactionClient, id: number) => {
   const notification = await prisma.notifications.delete({ where: { id } });
   return notification;
 };