+++ book-exchange-backend-mvc/src/models/payments.services.ts
@@ -1,21 +1,22 @@
 import { PrismaClient } from "@prisma/client";
+import { PrismaTransactionClient } from "@src/types/utils/primaAlias.types";
 import { CreatePaymentRequest, UpdatePaymentRequest } from "@src/types";
 
 export const createPayment = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   data: CreatePaymentRequest
 ) => {
   const payment = await prisma.payments.create({ data });
   return payment;
 };
 
-export const getPaymentById = async (prisma: PrismaClient, id: number) => {
+export const getPaymentById = async (prisma: PrismaTransactionClient, id: number) => {
   const payment = await prisma.payments.findUnique({ where: { id } });
   return payment;
 };
 
 export const updatePayment = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   id: number,
   data: UpdatePaymentRequest
 ) => {
@@ -23,7 +24,7 @@ export const updatePayment = async (
   return payment;
 };
 
-export const deletePayment = async (prisma: PrismaClient, id: number) => {
+export const deletePayment = async (prisma: PrismaTransactionClient, id: number) => {
   const payment = await prisma.payments.delete({ where: { id } });
   return payment;
 };