+++ book-exchange-backend-mvc/src/models/transactions.serviecs.ts
@@ -1,22 +1,23 @@
 import { PrismaClient } from "@prisma/client";
+import { PrismaTransactionClient } from "@src/types/utils/primaAlias.types";
 import { CreateTransactionRequest, UpdateTransactionRequest } from "@src/types";
 
 export const createTransaction = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   data: CreateTransactionRequest
 ) => {
   const transaction = await prisma.transactions.create({ data });
   return transaction;
 };
 
-export const getTransacitonById = async (prisma: PrismaClient, id: number) => {
+export const getTransacitonById = async (prisma: PrismaTransactionClient, id: number) => {
   const transaction = await prisma.transactions.findUnique({ where: { id } });
 
   return transaction;
 };
 
 export const updateTransaction = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   id: number,
   data: UpdateTransactionRequest
 ) => {
@@ -24,7 +25,7 @@ export const updateTransaction = async (
   return transaction;
 };
 
-export const deleteTransaciton = async (prisma: PrismaClient, id: number) => {
+export const deleteTransaciton = async (prisma: PrismaTransactionClient, id: number) => {
   const transaction = await prisma.transactions.delete({ where: { id } });
   return transaction;
 };