+++ book-exchange-backend-mvc/src/models/userAddress.services.ts
@@ -1,21 +1,22 @@
 import { PrismaClient } from "@prisma/client";
+import { PrismaTransactionClient } from "@src/types/utils/primaAlias.types";
 import { CreateUserAddressRequest, UpdateUserAddressRequest } from "@src/types";
 
 export const createUserAdderss = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   data: CreateUserAddressRequest
 ) => {
   const userAddress = await prisma.userAddress.create({ data });
   return userAddress;
 };
 
-export const getUserAddressById = async (prisma: PrismaClient, id: number) => {
+export const getUserAddressById = async (prisma: PrismaTransactionClient, id: number) => {
   const userAddress = await prisma.userAddress.findUnique({ where: { id } });
   return userAddress;
 };
 
 export const updateUserAddress = async (
-  prisma: PrismaClient,
+  prisma: PrismaTransactionClient,
   id: number,
   data: UpdateUserAddressRequest
 ) => {
@@ -23,7 +24,7 @@ export const updateUserAddress = async (
   return userAddress;
 };
 
-export const deleteUserAddress = async (prisma: PrismaClient, id: number) => {
+export const deleteUserAddress = async (prisma: PrismaTransactionClient, id: number) => {
   const userAddress = await prisma.userAddress.delete({ where: { id } });
   return userAddress;
 };