+++ book-exchange-backend-mvc/src/types/base/userProfilePictures.types.ts
@@ -10,6 +10,7 @@ export interface UserProfilePicture {
 export interface CreateUserProfilePictureRequest {
   userId: number;
   name: string;
+  isActive : boolean
 }
 
 export interface UpdateUserProfilePictureRequest {